requirements.txt
tweepy
python-dotenv
Pillow
requests
